+++ cloudhouse.py
@@ -14,20 +14,33 @@ claudemodels = ['claude-3-sonnet-20240229','claude-3-haiku-20240307','claude-3-o
 
 async def cloudhouse(user, message, replyto=None):
     bestmodel = False
-    chathistory.append(f'{user}: {message}')
-    print(f'debug chathistory: {chathistory}')
+    forcesubject = None
+    
+
     # select random member of cloudhouse
     #can call a specific friend with --friendname
-    if '--' in message:
+    if message.startswith('--'):
         friend = message.split()[0][2:]
         #strip the friendname from the message
         message = ' '.join(message.split()[1:])
         print(f'calling a specific friend: {friend}')
-        bestmodel = True # we want to always call the best model when calling a specific friend
+        bestmodel = True
     else:
         friend = random.choice(members)
+    if message.startswith('&&'):
+        #specific user
+        friend = message.split()[0][2:]
+        #strip the friendname from the message
+        message = ' '.join(message.split()[1:])
+        print(f'calling a specific friend: {friend}')
+        bestmodel = True
+        forcesubject = message
+        message = ''
+    else:
+        #add to chathistory
+        chathistory.append(f'{user}: {message}')
+        print(f'debug chathistory: {chathistory}')
     # get webhook from either ari_webhooks or personality.pwhl
-    print(f'debug friend: {friend}')
     all_webhooks = {**ari_webhooks.webhook_library, **personality.pwhl}
     try:
         webhook = all_webhooks[friend]
@@ -43,11 +56,14 @@ async def cloudhouse(user, message, replyto=None):
         print(f'error getting description for {username}')
         description = ''
     prompt = f'{personality.cloudhouse_prompt} information about your character and personality, {username} : {description}'
+    if forcesubject:
+        prompt = f'{prompt} subject of the message: {forcesubject}'
     print(prompt)
     # get message
     if friend in mainchars:
         if not bestmodel:
             cmodel = random.choice(claudemodels)
+            print(f'using random claudemodel: {cmodel}')
         else:
             print('using the best model as specific friend was requested')
             cmodel = 'claude-3-opus-20240229'
@@ -62,7 +78,7 @@ async def cloudhouse(user, message, replyto=None):
     if re.match(r'^\w+:',chresponse):
         chresponse = chresponse.split(':',1)[1]
 
-    chathistory.append(f'{username}: {message}')
+    chathistory.append(f'{username}: {chresponse}')
 
     return {'webhook':webhook,'message':chresponse}